Removed MLJ dependency and substituted with DecisionTree
