+++ test/orca_compression.jl
@@ -1,41 +1,46 @@
 using Test
 using SoleModels
 using SolePostHoc
-using MLJ
+using DecisionTree  
 using DataFrames
 using Random
 
 @testset "ORCA Genetic Compression Tests" begin
 
-    Xc, yc = @load_iris
-    Xc = DataFrame(Xc)
-    
+Random.seed!(42)
 
-    Random.seed!(42)
-    ttpairs = MLJ.MLJBase.train_test_pairs(Holdout(; shuffle = true), 1:length(yc), yc)
-    train_idx = ttpairs[1][1]
-    val_idx   = ttpairs[1][2]
+    n_samples = 100
+    n_features = 4
+    X_matrix = rand(n_samples, n_features)
+    
+    classlabels = ["classA", "classB", "classC"]
+    
+    yc_int = rand(1:length(classlabels), n_samples)
+    
+    yc_string = [classlabels[i] for i in yc_int]
+
+    Xc = DataFrame(X_matrix, :auto)
+    featurenames = string.(names(Xc))
+
+    train_idx = 1:80
+    val_idx   = 81:100
     
     n_trees_original = 10
-    DTModel = MLJ.@load RandomForestClassifier pkg=DecisionTree verbosity=0
-    model = DTModel(n_trees = n_trees_original)
-    mach = machine(model, Xc, yc)
-    MLJ.fit!(mach, rows = train_idx, verbosity = 0)
+    n_subfeatures = 2
     
-    featurenames = MLJ.report(mach).features
-    classlabels = mach.fitresult[2][sortperm((mach).fitresult[3])]
-    original_forest = solemodel(MLJ.fitted_params(mach).forest; featurenames, classlabels)
+    raw_forest = DecisionTree.build_forest(yc_int[train_idx], X_matrix[train_idx, :], n_subfeatures, n_trees_original)
+    
+    original_forest = solemodel(raw_forest; featurenames=featurenames, classlabels=classlabels)
 
-    f_val = Matrix{Float64}(Xc[val_idx, :])
-    l_val = string.(yc[val_idx])
+    f_val = X_matrix[val_idx, :]
+    l_val = yc_string[val_idx] 
 
     @test original_forest isa DecisionEnsemble
     @test length(original_forest.models) == n_trees_original
 
 
-    
     @testset "Mode: :size (Tree Selection)" begin
-        compressed_f = SolePostHoc.compression(
+        compressed_f = SolePostHoc.Orca.compression(
             original_forest, :size, f_val, l_val;
             population_size=10, n_generations=5
         )
@@ -45,7 +50,7 @@ using Random
     end
 
     @testset "Mode: :depth (Tree Pruning)" begin
-        compressed_f = SolePostHoc.compression(
+        compressed_f = SolePostHoc.Orca.compression(
             original_forest, :depth, f_val, l_val;
             population_size=10, n_generations=5
         )
@@ -55,7 +60,7 @@ using Random
     end
 
     @testset "Mode: :full_dimensional (Full Pipeline)" begin
-        compressed_f = SolePostHoc.compression(
+        compressed_f = SolePostHoc.Orca.compression(
             original_forest, :full_dimensional, f_val, l_val;
             population_size=10, n_generations=5
         )
@@ -65,7 +70,7 @@ using Random
     end
 
     @testset "Edge Cases and Errors" begin
-        @test_throws ErrorException SolePostHoc.compression(
+        @test_throws ErrorException SolePostHoc.Orca.compression(
             original_forest, :modalita_inesistente, f_val, l_val
         )
     end
